+++ tests/test_ambra.py
@@ -2,11 +2,11 @@ from ambra_sdk import API_VERSION, MODELS_VERSION, STORAGE_VERSION
 
 
 def test_api_version():
-    assert API_VERSION == 'LBL0022 v41.0 2020-10-07'
+    assert API_VERSION == 'LBL0022 v42.0 2020-12-02'
 
 
 def test_models_version():
-    assert MODELS_VERSION == 'LBL0022 v41.0 2020-10-07'
+    assert MODELS_VERSION == 'LBL0022 v42.0 2020-12-02'
 
 
 def test_storage_version():